Restructured utility functions into shared, seperate folder.
